--- lib/src/utils/git/mod.rs
@@ -1,4 +1,0 @@
-pub mod version;
-pub use version::*;
-pub mod repo;
-pub use repo::*;
