--- lib/src/utils/git/repo/mod.rs
@@ -1,10 +1,0 @@
-pub mod check_repo;
-pub use check_repo::check_repo;
-pub mod clone_repo;
-pub use clone_repo::clone_repo;
-pub mod clone_component;
-pub use clone_component::clone_component;
-pub mod checkout_latest_release;
-pub use checkout_latest_release::checkout_latest_release;
-pub mod fetch_tags;
-pub use fetch_tags::fetch_tags;
