--- lib/src/utils/git/version/mod.rs
@@ -1,8 +1,0 @@
-pub mod check_installed_version;
-pub use check_installed_version::check_installed_version;
-pub mod check_latest_version;
-pub use check_latest_version::check_latest_version;
-pub mod is_bin_installed;
-pub use is_bin_installed::is_bin_installed;
-pub mod get_bin_path;
-pub use get_bin_path::get_bin_path;
